Start to refactor towards a select loop.
This change splits out the command processing into a separate function and introduces a select loop.  The select really does nothing at the moment as it is just selecting on STDIN with no timeout.  Next thing to do I think is get the queue working such that q only adds things to the q and d sends the queue to SD.

After that we need to find a way to use callbacks to send things to SD a bit at a time.  This should help responsiveness considerably.